Refactors, patch support for visualDiff
Refactors out CameraVisPatch and material creation from CameraVis.
Supports visual differences based on restored/new patches.
